+++ src/testing/mod.rs
@@ -2,10 +2,12 @@
 
 mod simulated_user;
 mod test_data;
+mod test_drivers;
 mod test_signatures_building_coordinator;
 mod test_signing_drivers_context;
 
 pub use simulated_user::*;
 pub use test_data::*;
+pub use test_drivers::*;
 pub use test_signatures_building_coordinator::*;
 pub use test_signing_drivers_context::*;
