+++ src/testing/test_drivers/mod.rs
@@ -1,0 +1,7 @@
+mod test_parallel_batch_signing_driver;
+mod test_serial_batch_signing_driver;
+mod test_serial_single_signing_driver;
+
+pub use test_parallel_batch_signing_driver::*;
+pub use test_serial_batch_signing_driver::*;
+pub use test_serial_single_signing_driver::*;
