+++ src/testing/test_signatures_building_coordinator.rs
@@ -19,12 +19,23 @@ impl FactorResultsBuildingCoordinator {
         Self::new_test(
             all_factor_sources_in_profile,
             transactions,
-            SimulatedUser::prudent(),
+            SimulatedUser::prudent_no_fail(),
         )
     }
 
     pub fn test_prudent(transactions: impl IntoIterator<Item = TransactionIntent>) -> Self {
         Self::test_prudent_with_factors(FactorSource::all(), transactions)
+    }
+
+    pub fn test_prudent_with_retry(
+        transactions: impl IntoIterator<Item = TransactionIntent>,
+        retry: SimulatedUserRetries,
+    ) -> Self {
+        Self::new_test(
+            FactorSource::all(),
+            transactions,
+            SimulatedUser::prudent_with_retry(retry),
+        )
     }
 
     pub fn test_lazy_sign_minimum_no_retry_with_factors(
